todo-mvc: Rename action modules
