+++ todo-mvc/src/actions/widgetStoreActions.ts
